[FE] refactor: Search Artist
Search Artist 파일 분리 및
인터페이스 통일
