+++ frontend/pages/search/artists/index.tsx
@@ -4,9 +4,9 @@ import {
   StyledSearchArtistPage,
   StyledResult,
   StyledResultText,
-  StyledSearchArtistCards,
+  StyledSearchArtistList,
 } from "./styled";
-import SearchArtistCards from "../../../components/SearchSamples/artists";
+import SearchArtistList from "../../../components/SearchSamples/SearchArtistList";
 
 const SearchArtistPage = ({ filter }: { filter: string }): React.ReactElement => {
   const [sampleArtists, setSampleArtists] = useState([]);
@@ -23,9 +23,9 @@ const SearchArtistPage = ({ filter }: { filter: string }): React.ReactElement =>
       <StyledResult>
         <StyledResultText>{`'${filter}'의 검색 결과`}</StyledResultText>
       </StyledResult>
-      <StyledSearchArtistCards>
-        <SearchArtistCards data={sampleArtists} />
-      </StyledSearchArtistCards>
+      <StyledSearchArtistList>
+        <SearchArtistList data={sampleArtists} />
+      </StyledSearchArtistList>
     </StyledSearchArtistPage>
   );
 };
